Move project changed listener class to event_listener package
