+++ src/main/java/jp/ex_t/kazuaki/change_vision/PairModeling.kt
@@ -9,6 +9,7 @@
 package jp.ex_t.kazuaki.change_vision
 
 import com.change_vision.jude.api.inf.AstahAPI
+import jp.ex_t.kazuaki.change_vision.event_listener.ProjectChangedListener
 
 class PairModeling(topic: String, private val clientId: String, private val brokerAddress: String) {
     private val topicTransaction = "$topic/transaction"
