+++ src/main/java/jp/ex_t/kazuaki/change_vision/ReflectTransaction.kt
@@ -14,6 +14,7 @@ import com.change_vision.jude.api.inf.project.ProjectAccessor
 import com.change_vision.jude.api.inf.ui.IPluginActionDelegate.UnExpectedException
 import jp.ex_t.kazuaki.change_vision.apply_transaction.ClassDiagramApplyTransaction
 import jp.ex_t.kazuaki.change_vision.apply_transaction.MindmapDiagramApplyTransaction
+import jp.ex_t.kazuaki.change_vision.event_listener.ProjectChangedListener
 import javax.swing.SwingUtilities
 
 class ReflectTransaction(private val projectChangedListener: ProjectChangedListener) {
